# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Origin travel API",
    "version": "0.1.0",
    "description": "Travel search and supplier booking. AI tools can call these routes later."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/v1/sokosumi/tick": {
      "post": {
        "summary": "Check assigned Coworker tasks (execution rehearsal)",
        "tags": [
          "Sokosumi"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          }
        },
        "security": [],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/flights/search": {
      "post": {
        "summary": "Search flight offers",
        "tags": [
          "Flights"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "slices": {
                    "minItems": 1,
                    "maxItems": 6,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "origin": {
                          "type": "string",
                          "pattern": "^[A-Z]{3}$"
                        },
                        "destination": {
                          "type": "string",
                          "pattern": "^[A-Z]{3}$"
                        },
                        "departure_date": {
                          "type": "string",
                          "format": "date",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
                        }
                      },
                      "required": [
                        "origin",
                        "destination",
                        "departure_date"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "passengers": {
                    "minItems": 1,
                    "maxItems": 9,
                    "type": "array",
                    "items": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "enum": [
                                "adult"
                              ]
                            }
                          },
                          "required": [
                            "type"
                          ],
                          "additionalProperties": false
                        },
                        {
                          "type": "object",
                          "properties": {
                            "age": {
                              "type": "integer",
                              "minimum": 0,
                              "maximum": 17
                            }
                          },
                          "required": [
                            "age"
                          ],
                          "additionalProperties": false
                        }
                      ]
                    }
                  },
                  "cabin_class": {
                    "default": "economy",
                    "type": "string",
                    "enum": [
                      "economy",
                      "premium_economy",
                      "business",
                      "first"
                    ]
                  },
                  "max_connections": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 3
                  }
                },
                "required": [
                  "slices",
                  "passengers"
                ],
                "additionalProperties": false
              },
              "example": {
                "slices": [
                  {
                    "origin": "SIN",
                    "destination": "BKK",
                    "departure_date": "2026-11-05"
                  }
                ],
                "passengers": [
                  {
                    "type": "adult"
                  }
                ],
                "cabin_class": "economy",
                "max_connections": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          },
          "400": {
            "description": "Invalid input, including past dates or check-out on/before check-in.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Request or supplier rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "502": {
            "description": "Supplier rejected the request or returned an invalid response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Supplier credentials are not configured.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "504": {
            "description": "Supplier did not respond before the timeout.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        },
        "description": "Search Singapore to Bangkok with future travel dates. The example dates refresh daily. Results come from the configured supplier; provider and environment identify production or sandbox data. Availability and prices can change."
      }
    },
    "/v1/flights/offers/{id}": {
      "get": {
        "summary": "Refresh a flight offer",
        "tags": [
          "Flights"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,100}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/flights/bookings": {
      "post": {
        "summary": "Book a refreshed flight offer",
        "tags": [
          "Flights"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "offer_id": {
                    "type": "string",
                    "pattern": "^[a-zA-Z0-9_-]{1,100}$"
                  },
                  "passengers": {
                    "minItems": 1,
                    "maxItems": 9,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "given_name": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 200
                        },
                        "family_name": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 200
                        },
                        "id": {
                          "type": "string",
                          "pattern": "^[a-zA-Z0-9_-]{1,100}$"
                        },
                        "born_on": {
                          "type": "string",
                          "format": "date",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                        },
                        "phone_number": {
                          "type": "string",
                          "pattern": "^\\+[1-9]\\d{6,14}$"
                        },
                        "title": {
                          "type": "string",
                          "enum": [
                            "mr",
                            "ms",
                            "mrs",
                            "miss",
                            "dr"
                          ]
                        },
                        "gender": {
                          "type": "string",
                          "enum": [
                            "m",
                            "f"
                          ]
                        },
                        "infant_passenger_id": {
                          "type": "string",
                          "pattern": "^[a-zA-Z0-9_-]{1,100}$"
                        },
                        "identity_documents": {
                          "maxItems": 5,
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "type": {
                                "type": "string",
                                "enum": [
                                  "passport",
                                  "known_traveler_number",
                                  "passenger_redress_number"
                                ]
                              },
                              "unique_identifier": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 200
                              },
                              "issuing_country_code": {
                                "type": "string",
                                "pattern": "^[A-Z]{2}$"
                              },
                              "expires_on": {
                                "type": "string",
                                "format": "date",
                                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
                              }
                            },
                            "required": [
                              "type",
                              "unique_identifier"
                            ],
                            "additionalProperties": false
                          }
                        }
                      },
                      "required": [
                        "given_name",
                        "family_name",
                        "id",
                        "born_on",
                        "email",
                        "phone_number",
                        "title",
                        "gender"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "max_total": {
                    "type": "object",
                    "properties": {
                      "amount": {
                        "type": "string",
                        "pattern": "^(0|[1-9]\\d{0,12})(\\.\\d{1,6})?$"
                      },
                      "currency": {
                        "type": "string",
                        "pattern": "^[A-Z]{3}$"
                      }
                    },
                    "required": [
                      "amount",
                      "currency"
                    ],
                    "additionalProperties": false
                  },
                  "confirm": {
                    "type": "boolean",
                    "enum": [
                      true
                    ]
                  }
                },
                "required": [
                  "offer_id",
                  "passengers",
                  "max_total",
                  "confirm"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{8,128}$"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/stays/hotels/{id}": {
      "get": {
        "summary": "Get hotel photos, facilities, and review highlights",
        "tags": [
          "Stays"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,100}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/stays/bookings": {
      "post": {
        "summary": "Book a hotel offer within a budget",
        "tags": [
          "Stays"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "offer_id": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9_=-]{20,4000}$",
                    "description": "offer_id from /v1/stays/search"
                  },
                  "guests": {
                    "minItems": 1,
                    "maxItems": 10,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "given_name": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 200
                        },
                        "family_name": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 200
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                        },
                        "occupancy_number": {
                          "default": 1,
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 5
                        }
                      },
                      "required": [
                        "given_name",
                        "family_name",
                        "email"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "max_total": {
                    "type": "object",
                    "properties": {
                      "amount": {
                        "type": "string",
                        "pattern": "^(0|[1-9]\\d{0,12})(\\.\\d{1,6})?$"
                      },
                      "currency": {
                        "type": "string",
                        "pattern": "^[A-Z]{3}$"
                      }
                    },
                    "required": [
                      "amount",
                      "currency"
                    ],
                    "additionalProperties": false
                  },
                  "confirm": {
                    "type": "boolean",
                    "enum": [
                      true
                    ]
                  }
                },
                "required": [
                  "offer_id",
                  "guests",
                  "max_total",
                  "confirm"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{8,128}$"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/bookings/stay/{id}": {
      "get": {
        "summary": "Get hotel booking status",
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,100}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/bookings/flight/{id}": {
      "get": {
        "summary": "Get flight booking status",
        "tags": [
          "Bookings"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,100}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}